+++ checkpoint/orbax/checkpoint/experimental/v1/_src/serialization/compatibility.py
@@ -20,6 +20,7 @@ from absl import logging
 from etils import epath
 import jax
 from jax import tree_util as jtu
+import numpy as np
 from orbax.checkpoint._src.futures import future
 from orbax.checkpoint._src.metadata import sharding as sharding_metadata
 from orbax.checkpoint._src.metadata import value as value_metadata
@@ -28,6 +29,8 @@ from orbax.checkpoint._src.serialization import types as types_v0
 from orbax.checkpoint.experimental.v1._src.context import context as context_lib
 from orbax.checkpoint.experimental.v1._src.path import types as path_types
 from orbax.checkpoint.experimental.v1._src.serialization import array_leaf_handler
+from orbax.checkpoint.experimental.v1._src.serialization import numpy_leaf_handler
+from orbax.checkpoint.experimental.v1._src.serialization import scalar_leaf_handler
 from orbax.checkpoint.experimental.v1._src.serialization import types
 from orbax.checkpoint.experimental.v1._src.synchronization import synchronization
 from orbax.checkpoint.experimental.v1._src.tree import types as tree_types
@@ -97,13 +100,31 @@ def _construct_serialization_context(
 def _construct_deserialization_param(
     info: types_v0.ParamInfo,
     restore_args: types_v0.RestoreArgs,
-) -> types.DeserializationParam[array_leaf_handler.AbstractArray | None]:
+) -> types.DeserializationParam[
+    array_leaf_handler.AbstractArray
+    | numpy_leaf_handler.AbstractNumpy
+    | scalar_leaf_handler.AbstractScalar
+    | None
+]:
   """Constructs a DeserializationParam from a ParamInfo and RestoreArg."""
+
+  logging.vlog(1, 'compatibility.py: restore_args: %s', restore_args)
 
   value = None
 
-  if isinstance(restore_args, type_handlers_v0.ArrayRestoreArgs):
-    # Array type, construct value as jax.ShapeDtypeStruct.
+  if restore_args.restore_type == np.ndarray:
+    # Numpy type
+    value = numpy_leaf_handler.NumpyShapeDtype(
+        dtype=restore_args.dtype,
+        shape=None,
+    )
+  elif restore_args.restore_type in (scalar_leaf_handler.Scalar.__args__):
+    # JAX Array type, construct value as jax.ShapeDtypeStruct.
+    value = scalar_leaf_handler.AbstractScalar(
+        dtype=restore_args.restore_type,
+    )
+  elif isinstance(restore_args, type_handlers_v0.ArrayRestoreArgs):
+    # JAX Array type, construct value as jax.ShapeDtypeStruct.
     arg = cast(type_handlers_v0.ArrayRestoreArgs, restore_args)
 
     logging.info(
@@ -211,7 +232,27 @@ def _convert_v1_metadata_to_v0(
         dtype=metadata.dtype,
         storage=metadata.storage_metadata,
     )
-    logging.info('ArrayMetadata: %s', ret)
+    logging.vlog(1, 'ArrayMetadata: %s', ret)
+    return ret
+  elif isinstance(metadata, numpy_leaf_handler.NumpyMetadata):
+    metadata = cast(numpy_leaf_handler.NumpyMetadata, metadata)
+    ret = value_metadata.ArrayMetadata(
+        name=name,
+        directory=directory,
+        sharding=None,
+        shape=metadata.shape,
+        dtype=metadata.dtype,
+        storage=metadata.storage_metadata,
+    )
+    logging.vlog(1, 'NumpyMetadata: %s', ret)
+    return ret
+  elif isinstance(metadata, scalar_leaf_handler.AbstractScalar):
+    ret = value_metadata.ScalarMetadata(
+        name=name,
+        directory=directory,
+        dtype=metadata.dtype,
+    )
+    logging.vlog(1, 'ScalarMetadata: %s', ret)
     return ret
   else:
     logging.warning(
@@ -281,7 +322,7 @@ class CompatibleTypeHandler(
 
     params = []
     if args is None:
-      args = [types_v0.RestoreArgs(restore_type=types.Leaf)] * len(infos)
+      args = [types_v0.RestoreArgs()] * len(infos)
 
     for info, restore_arg in zip(infos, args):
       if logging.vlog_is_on(1):
@@ -331,7 +372,7 @@ class CompatibleTypeHandler(
   def memory_size(
       self, values: Sequence[types.Leaf]
   ) -> Sequence[Tuple[int, int]]:
-    # this only works for leaf handler that based on V0 ArrayHandler and stored
+    # this only works for leaf handler that based on V0 TypeHandlers and stored
     # it in self._leaf_handler._handler_impl.
     if hasattr(self._leaf_handler, '_handler_impl'):
       v0_handler = self._leaf_handler._handler_impl  # pylint: disable=protected-access
@@ -391,4 +432,22 @@ def get_compatible_type_handler_registry(
       ),
       override=True,
   )
+  type_handler_registry.add(
+      np.ndarray,
+      CompatibleTypeHandler(
+          numpy_leaf_handler.NumpyLeafHandler(context=context),
+          typestr='np.ndarray',
+      ),
+      override=True,
+  )
+
+  for scalar_type in (int, float, bytes, np.number):
+    type_handler_registry.add(
+        scalar_type,
+        CompatibleTypeHandler(
+            scalar_leaf_handler.ScalarLeafHandler(context=context),
+            typestr='scalar',
+        ),
+        override=True,
+    )
   return type_handler_registry